[XrdCl] Allow loading XrdEc default plug-in based on cfg file.
