+++ src/XrdCl/XrdClEcHandler.hh
@@ -36,17 +36,18 @@ namespace XrdCl
       {
       }
 
-      XrdCl::XRootDStatus Open( uint16_t                  flags,
-                                XrdCl::ResponseHandler   *handler,
-                                uint16_t                  timeout )
+      XRootDStatus Open( uint16_t                  flags,
+                         XrdCl::ResponseHandler   *handler,
+                         uint16_t                  timeout )
       {
+        // TODO if plgr is empty issue locate to figure out the plgr
+
         if( ( flags & XrdCl::OpenFlags::Write ) || ( flags & XrdCl::OpenFlags::Update ) )
         {
           if( !( flags & XrdCl::OpenFlags::New )    || // it has to be a new file
                ( flags & XrdCl::OpenFlags::Delete ) || // truncation is not supported
                ( flags & XrdCl::OpenFlags::Read ) )    // write + read is not supported
-            return XrdCl::XRootDStatus( XrdCl::stError, XrdCl::errNotSupported );
-    
+            return XrdCl::XRootDStatus( XrdCl::stError, XrdCl::errNotSupported ); // TODO call handler
           writer.reset( new XrdEc::StrmWriter( *objcfg ) );
           writer->Open( handler, timeout );
           return XrdCl::XRootDStatus();
@@ -55,15 +56,26 @@ namespace XrdCl
         if( flags & XrdCl::OpenFlags::Read )
         {
           if( flags & XrdCl::OpenFlags::Write )
-            return XrdCl::XRootDStatus( XrdCl::stError, XrdCl::errNotSupported );
+            return XrdCl::XRootDStatus( XrdCl::stError, XrdCl::errNotSupported ); // TODO call handler
     
           reader.reset( new XrdEc::Reader( *objcfg ) );
           reader->Open( handler, timeout );
           return XrdCl::XRootDStatus(); 
         }
-    
-        return XrdCl::XRootDStatus( XrdCl::stError, XrdCl::errNotSupported );
+
+        return XrdCl::XRootDStatus( XrdCl::stError, XrdCl::errNotSupported ); // TODO call handler
       }
+
+      XRootDStatus Open( const std::string &url,
+                                 OpenFlags::Flags   flags,
+                                 Access::Mode       mode,
+                                 ResponseHandler   *handler,
+                                 uint16_t           timeout )
+      {
+        (void)url; (void)mode;
+        return Open( flags, handler, timeout );
+      }
+
 
       //------------------------------------------------------------------------
       //!
@@ -180,85 +192,56 @@ namespace XrdCl
 
   };
 
-  EcHandler* GetEcHandler( const URL &headnode, const URL &redirurl )
+  //----------------------------------------------------------------------------
+  //! Plugin factory
+  //----------------------------------------------------------------------------
+  class EcPlugInFactory : public PlugInFactory
   {
-    const URL::ParamsMap &params = redirurl.GetParams();
-    // make sure all the xrdec. tokens are present and the values are sane
-    URL::ParamsMap::const_iterator itr = params.find( "xrdec.nbdta" );
-    if( itr == params.end() ) return nullptr;
-    uint8_t nbdta = std::stoul( itr->second );
+    public:
+      //------------------------------------------------------------------------
+      //! Constructor
+      //------------------------------------------------------------------------
+      EcPlugInFactory( uint8_t nbdta, uint8_t nbprt, uint64_t chsz,
+                       std::vector<std::string> && plgr ) :
+        nbdta( nbdta ), nbprt( nbprt ), chsz( chsz ), plgr( std::move( plgr ) )
+      {
+      }
 
-    itr = params.find( "xrdec.nbprt" );
-    if( itr == params.end() ) return nullptr;
-    uint8_t nbprt = std::stoul( itr->second );
+      //------------------------------------------------------------------------
+      //! Destructor
+      //------------------------------------------------------------------------
+      virtual ~EcPlugInFactory()
+      {
+      }
 
-    itr = params.find( "xrdec.blksz" );
-    if( itr == params.end() ) return nullptr;
-    uint64_t blksz = std::stoul( itr->second );
+      //------------------------------------------------------------------------
+      //! Create a file plug-in for the given URL
+      //------------------------------------------------------------------------
+      virtual FilePlugIn *CreateFile( const std::string &u )
+      {
+        URL url( u );
+        XrdEc::ObjCfg *objcfg = new XrdEc::ObjCfg( url.GetPath(), nbdta, nbprt,
+                                                   chsz, false, true );
+        objcfg->plgr = std::move( plgr );
+        return new EcHandler( url, objcfg, nullptr );
+      }
 
-    itr = params.find( "xrdec.plgr" );
-    if( itr == params.end() ) return nullptr;
-    std::vector<std::string> plgr;
-    Utils::splitString( plgr, itr->second, "," );
-    if( plgr.size() < nbdta + nbprt ) return nullptr;
+      //------------------------------------------------------------------------
+      //! Create a file system plug-in for the given URL
+      //------------------------------------------------------------------------
+      virtual FileSystemPlugIn *CreateFileSystem( const std::string &url )
+      {
+        return nullptr;
+      }
 
-    itr = params.find( "xrdec.objid" );
-    if( itr == params.end() ) return nullptr;
-    std::string objid = itr->second;
+    private:
+      uint8_t  nbdta;
+      uint8_t  nbprt;
+      uint64_t chsz;
+      std::vector<std::string> plgr;
+  };
 
-    itr = params.find( "xrdec.format" );
-    if( itr == params.end() ) return nullptr;
-    size_t format = std::stoul( itr->second );
-    if( format != 1 ) return nullptr; // TODO use constant
-
-    std::vector<std::string> dtacgi;
-    itr = params.find( "xrdec.dtacgi" );
-    if( itr != params.end() )
-    {
-      Utils::splitString( dtacgi, itr->second, "," );
-      if( plgr.size() != dtacgi.size() ) return nullptr;
-    }
-
-    std::vector<std::string> mdtacgi;
-    itr = params.find( "xrdec.mdtacgi" );
-    if( itr != params.end() ) 
-    {
-      Utils::splitString( mdtacgi, itr->second, "," );
-      if( plgr.size() != mdtacgi.size() ) return nullptr;
-    }
-
-    itr = params.find( "xrdec.cosc" );
-    if( itr == params.end() ) return nullptr;
-    std::string cosc_str = itr->second;
-    if( cosc_str != "true" && cosc_str != "false" ) return nullptr;
-    bool cosc = cosc_str == "true";
-
-    itr = params.find( "xrdec.cksum" );
-    if( cosc && itr == params.end() ) return nullptr;
-    std::string ckstype = itr->second;
-
-    std::string chdigest;
-    itr = params.find( "xrdec.chdigest" );
-    if( itr == params.end() )
-      chdigest = "crc32c";
-    else
-      chdigest = itr->second;
-    bool usecrc32c = ( chdigest == "crc32c" );
-
-    XrdEc::ObjCfg *objcfg = new XrdEc::ObjCfg( objid, nbdta, nbprt, blksz / nbdta, usecrc32c );
-    objcfg->plgr    = std::move( plgr );
-    objcfg->dtacgi  = std::move( dtacgi );
-    objcfg->mdtacgi = std::move( mdtacgi );
-
-    std::unique_ptr<CheckSumHelper> cksHelper( cosc ? new CheckSumHelper( "", ckstype ) : nullptr );
-    if( cksHelper )
-    {
-      auto st = cksHelper->Initialize();
-      if( !st.IsOK() ) return nullptr;
-    }
-
-    return new EcHandler( headnode, objcfg, std::move( cksHelper ) );
-  }
+  EcHandler* GetEcHandler( const URL &headnode, const URL &redirurl );
 
 } /* namespace XrdCl */
